requirements.txt
python-telegram-bot==20.3
transformers==4.34.0
torch==2.0.1